# perf(webidl): optimize createRecordConverter()