--- a/ext/webidl/00_webidl.js
+++ b/ext/webidl/00_webidl.js
@@ -848,23 +848,23 @@
       }
       const result = {};
       const isProxy = core.isProxy(V);
-      if (isProxy) {
-        const keys = ReflectOwnKeys(V);
-        for (const key of keys) {
-          const desc = ObjectGetOwnPropertyDescriptor(V, key);
-          if (desc !== undefined && desc.enumerable === true) {
-            const typedKey = keyConverter(key, opts);
-            const value = V[key];
-            const typedValue = valueConverter(value, opts);
-            result[typedKey] = typedValue;
-          }
-        }
-      } else {
-        // Fast path for common case (not a Proxy)
+      // Fast path for common case (not a Proxy)
+      if (!isProxy) {
         for (const key in V) {
           if (!ObjectPrototypeHasOwnProperty(V, key)) {
             continue;
           }
+          const typedKey = keyConverter(key, opts);
+          const value = V[key];
+          const typedValue = valueConverter(value, opts);
+          result[typedKey] = typedValue;
+        }
+      }
+      // Slow path if Proxy (e.g: in WPT tests)
+      const keys = ReflectOwnKeys(V);
+      for (const key of keys) {
+        const desc = ObjectGetOwnPropertyDescriptor(V, key);
+        if (desc !== undefined && desc.enumerable === true) {
           const typedKey = keyConverter(key, opts);
           const value = V[key];
           const typedValue = valueConverter(value, opts);