Cuts self-time by ~6x, 172ns/iter => 22ns/iter benched on 1M Response builds / HeadersInit calls

This should be functionally identical to the old code since:
1. `for...in` loops on objects iterate over all enumberable string keys of an object (including inherited)
2. Then we filter out inherited keys with `.hasOwnProperty(...)`

**Note:** WebIDL / WPT is coupled to a subpar implementation so our hands are tied, unless an isProxy fast path check isn't too expensive

## Benchmarks

```
No fast path:
resp_w_bh:           	n = 1000000, dt = 1.295s, r = 772201/s, t = 1294ns/op
Fast path with isProxy check:
resp_w_bh:           	n = 1000000, dt = 1.146s, r = 872600/s, t = 1146ns/op
Forced fast path:
resp_w_bh:           	n = 1000000, dt = 1.112s, r = 899281/s, t = 1112ns/op
```

Pending on https://github.com/denoland/deno/pull/12288