@lucacasonato 2021-10-01
This is not equal. `for .. in` also enumerates prototype fields, while `ReflectOwnKeys` does not.