@lucacasonato 2021-10-01
Actually, NVM. Inherited keys are filtered out via `hasOwnProperty`